Add missing Stream overrides
It's important for perf to override the base implementations, in particular for Read/WriteByte.

I just copied the existing implementations to the new overrides and fixed up the implementations with as minimal changes as possible. I also tried to retain the style of the surrounding code.

And none of the existing implementations have tests, so I didn't add any for these, as much as it pained me :)